+++ tests/sys/conn_can/main.c
@@ -592,7 +592,7 @@ static int _can_handler(int argc, char **argv)
 
 static void *_receive_thread(void *args)
 {
-    int thread_nb = (int)args;
+    int thread_nb = (intptr_t)args;
     struct can_frame frame;
     msg_t msg, msg_queue[RECEIVE_THREAD_MSG_QUEUE_SIZE];
 
@@ -729,7 +729,7 @@ static const shell_command_t _commands[] = {
 
 int main(void)
 {
-    for (int i = 0; i < RCV_THREAD_NUMOF; i++) {
+    for (intptr_t i = 0; i < RCV_THREAD_NUMOF; i++) {
         receive_pid[i] = thread_create(thread_stack[i], THREAD_STACKSIZE,
                                        THREAD_PRIORITY_MAIN - 1,
                                        THREAD_CREATE_STACKTEST, _receive_thread,
